Move telegram package to src and remove notifier package
